Make MPS internal and flatten directory structure
All MPS files (headers and compilation units) are now in library/*
